Move SchemaDefinitionCache to be accessible via VC api
